+++ Libplanet/Store/Trie/ITrie.cs
@@ -45,8 +45,7 @@ namespace Libplanet.Store.Trie
         /// <summary>
         /// Cleans up and stores the <see cref="ITrie"/> in storage.
         /// </summary>
-        /// <param name="rehearsal">Whether to store nodes.</param>
         /// <returns>Returns new committed <see cref="ITrie"/>.</returns>
-        ITrie Commit(bool rehearsal = false);
+        ITrie Commit();
     }
 }